Revert "Update pysnmp-lextudio version and dependencies"
This reverts commit 1d97495709119fe16fa006452f2a80bbf32291b4.

Reason for revert: this breaks virtualpdu which is essential for snmp
support in ironic. Unfortunately the migration to asyncio in
pysnmp-lextudio is more destructive than expected and we need to
revert to precedent version to be able to unlock ironic CI until
we migrate also the virtualpdu code to be compatible with asyncio.

Change-Id: Idf965c48a219c83250ba39c3f615a9ba8adcad26