add importlib-resources to replace pkg_resources usage
pkg_resources comes from setuptools which starting with Python 3.12 will
not be installed by default into virtualenvs which will break projects
implicitly assuming pkg_resources is available. Additionally setuptools
has deprecated pkg_resources so the best plan is to use the stdlib
provided solution, which is not available until Python 3.9 so we must
use the backport package for Python 3.8 support.

Change-Id: I15d3a4858383d7e0acc65e62b886c584bbc363ab