Add django-libsass and libsass
The django-pyscss and pyscss libraries currently being used in Horizon
are not currently maintained and have an increasing number of problems
that we have to work around. We want to instead switch to using the
libsass library, which is using a C library of the same name under the
hood, which in turn contains the official implementation of the sass and
scss parsers.

Change-Id: I6d5c895e36766c9b1e1bb41f4c98468ba1b51a64