Add netifaces-2
netifaces is dead, but openstacksdk depends on it. netifaces-2 is
a replacement for as much of it as openstacksdk uses. Add it here
so that we can unbreak sdk issues.

There is also a suggestion to replace with psutils, but the
corresponding psutils functionality doesn't actually do the same
thing.

Change-Id: Iade27d584791d6a4ea1d986e52412f266a99d14b