Remove os-traits from upper-constraints
Like many other libraries and external tools, os-traits need not be
globally synchronized in the same way other libraries are. A project
that has specific requirements can manage those requirements
themselves. The os-traits "library" is just a list of additive string
constants with no real expectation of ever removing something. Tracking
it in upper-constraints really just creates hoops for developers and
release team members that are not valuable or necessary.

Change-Id: Ic29c5994d985bdafefbcde85a3f9b6f62de7f056