Add testing tool validate-pyproject
As we are now managing pyproject.toml files for our packages, I think we
should have a test tool to validate the syntax of these files match the
specification.

Is the library actively maintained?

Yes, the last release of the main package was March 21st, 2025 with
commits two weeks ago.

Is the library good code?

Yes, it appears to be. It is build around JSON schemas to help with
validation.

Is the library license compatible?

Yes, it is MPL-2.0 like some of our other testing tools.

Is the library already packaged in the distros we target?

Yes, on Ubuntu it is python3-validate-pyproject

Is the function of this library already covered by other libraries in
global-requirements.txt?

Not that I am aware of.

Is the library required for OpenStack project or related dev or infrastructure
setup?

Octavia would like to add this to our pep8 checks.

Change-Id: I4a278fb098b9386d8e09d57fb332346f7da4910b